Rewrite kmermatcher and clustering, regression works

+++ src/commons/CSProfile.h
@@ -60,10 +60,10 @@ class CSProfile {
     float * sums;
 public:
 
-    CSProfile(size_t maxSeqLen) {
+    CSProfile(unsigned int maxSeqLen) {
         ctxLib = ContextLibrary::getContextLibraryInstance();
         this->profile = (float * )mem_align(ALIGN_FLOAT, (Sequence::PROFILE_AA_SIZE + 4) * maxSeqLen * sizeof(float));
-        int segmentSize = (maxSeqLen+VECSIZE_FLOAT-1)/VECSIZE_FLOAT;
+        unsigned int segmentSize = (maxSeqLen+VECSIZE_FLOAT-1)/VECSIZE_FLOAT;
         this->pp = (float * ) mem_align(ALIGN_FLOAT, 4000 * segmentSize * VECSIZE_FLOAT * sizeof(float));
         this->maximums = (float * ) mem_align(ALIGN_FLOAT,  segmentSize * VECSIZE_FLOAT * sizeof(float));
         this->sums = (float * ) mem_align(ALIGN_FLOAT,  segmentSize * VECSIZE_FLOAT * sizeof(float));
